requirements.txt
numpy>=1.21.0
opencv-python>=4.5.0
pillow>=9.0.0
scipy>=1.7.0
scikit-learn>=1.0.0
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.10.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.3.0
scikit-image>=0.19.0
tqdm>=4.62.0
jupyter>=1.0.0
albumentations>=1.3.0
pywavelets>=1.3.0
gradio>=4.0.0
